// devui/src/app/shared/models/dto/provisional-rule-dto.ts
import { Users} from '../users';
import { RuleInfo } from '../rule-info';
import { RuleReference } from '../rule-reference';
import { EclReference } from '../ecl-reference';
import { OpportunityValueDto } from '../opportunity-value';
import { RuleProcedureCodes } from '../rule-procedure-codes';
import { RuleImpactAnalysisRun } from '../rule-impact-analysis-run';
import { RuleReferenceUpdates } from '../rule-reference-updates';
import { RuleCodeDto } from './rule-code-dto';
import { RuleNoteTabDto } from './rule-notetab-dto';
export class ProvisionalRuleDto {
    ruleInfo: RuleInfo;
    procedureCodeDto:RuleCodeDto[];
    diagnosisCodeDto: RuleCodeDto[];
    lobs: number[];
    states: number[];
    jurisdictions: number[];
    procedureCodes: RuleProcedureCodes[];
    excludedSpecialityTypes: any[];
    excludedSubspecialityTypes: any[];
    includedBills: any[];
    excludedBills: any[];
    includedSpecialityTypes: any[];
    includedSubspecialityTypes: any[];
    user: number;
    action: string;
    newProvRule : boolean;
    fromMaintenanceProcess: boolean;
    selectedReviewStatus: string;
    ruleReviewComments: string;
    eclReference: EclReference[];
    ruleReference: RuleReference[];
    oppValueDto: OpportunityValueDto;
    assignedTo: number;
    ruleImpactAnalysisDetails: RuleImpactAnalysisRun;
    ruleRefUpdates: RuleReferenceUpdates[];
    claimTypes?: any;
    ruleNotesComments : RuleNoteTabDto;
    pdgTemplateDto: any;
    policyPackages: any[];
}